backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.4
Django==6.0.2
django-allauth==65.14.3
django-cors-headers==4.9.0
django-filter==24.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
Pillow==11.2.0
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
